SAVE SYSTEM
Blocks can now be saved, and then placed. The logic gates get simulated when the new block gets saved, that way the simulation only needs to happen once and not when the real simulation is running. This way unnecessary combinations might get simulated though. Editing existing blocks is still required.